requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
redis==5.0.1
langgraph==0.0.26
langchain==0.1.9
transformers==4.38.0
torch==2.2.0
accelerate==0.27.0
pydantic==2.6.1
aiofiles==23.2.1
numpy==1.26.4
pytest-asyncio
chromadb
Pillow
pyteseract
streamlit